add movies

+++ Cinema.Movie/Cinema.Movie.Web/Modules/Movie/Video/VideoColumns.cs
@@ -26,7 +26,7 @@ namespace Cinema.Movie.Movie.Columns
         public String Storyline { get; set; }
         public DateTime PlannePublishDate { get; set; }
         public DateTime ActualPublishDateTime { get; set; }
-        public Int16 MovieId { get; set; }
+        public Int64 MovieId { get; set; }
         public Int16 ServiceId { get; set; }
         public String ServiceName { get; set; }
     }
