+++ Cinema.Movie/Cinema.Movie.Web/Modules/Movie/Video/VideoForm.cs
@@ -23,7 +23,7 @@ namespace Cinema.Movie.Movie.Forms
         public String Storyline { get; set; }
         public DateTime PlannePublishDate { get; set; }
         public DateTime ActualPublishDateTime { get; set; }
-        public Int16 MovieId { get; set; }
+        public Int64 MovieId { get; set; }
         public Int16 ServiceId { get; set; }
     }
 }
